refactor: component & layout
